git status
