+++ lib/assets/javascripts/cartodb3/editor/style/style-form/style-properties-form/style-labels-properties-form-model.js
@@ -1,8 +1,9 @@
-var cdb = require('cartodb.js');
 var _ = require('underscore');
-var StyleFormComponents = require('../style-form-components-dictionary');
+var StyleFormDefaultModel = require('../style-form-default-model');
 
-module.exports = cdb.core.Model.extend({
+module.exports = StyleFormDefaultModel.extend({
+
+  _FORM_NAME: 'labels',
 
   parse: function (r) {
     return _.extend(
@@ -11,18 +12,6 @@ module.exports = cdb.core.Model.extend({
         overlap: r.overlap.toString()
       }
     );
-  },
-
-  initialize: function (attrs, opts) {
-    if (!opts.styleModel) throw new Error('Style model is required');
-    this._styleModel = opts.styleModel;
-    this._querySchemaModel = opts.querySchemaModel;
-    this.schema = this._generateSchema();
-    this._initBinds();
-  },
-
-  _initBinds: function () {
-    this.bind('change', this._onChange, this);
   },
 
   _onChange: function () {
@@ -40,17 +29,5 @@ module.exports = cdb.core.Model.extend({
     }
 
     this._styleModel.set('labels', labelsData);
-  },
-
-  _generateSchema: function () {
-    var self = this;
-    return _.reduce(this.attributes, function (memo, value, key) {
-      var d = StyleFormComponents['labels-' + key];
-      if (d) {
-        memo[key] = d(self._querySchemaModel, self._styleModel.get('type'));
-      }
-      return memo;
-    }, {});
   }
-
 });
