+++ lib/assets/javascripts/cartodb3/editor/style/style-form/style-properties-form/style-shape-properties-form-model.js
@@ -1,9 +1,8 @@
-var cdb = require('cartodb.js');
 var _ = require('underscore');
 var StylesFactory = require('../../styles-factory');
-var StyleFormComponents = require('../style-form-components-dictionary');
+var StyleFormDefaultModel = require('../style-form-default-model');
 
-module.exports = cdb.core.Model.extend({
+module.exports = StyleFormDefaultModel.extend({
 
   parse: function (r) {
     var geom = r.geom;
@@ -25,31 +24,8 @@ module.exports = cdb.core.Model.extend({
     return attrs;
   },
 
-  initialize: function (attrs, opts) {
-    if (!opts.styleModel) throw new Error('Style model is required');
-    this._styleModel = opts.styleModel;
-    this._querySchemaModel = opts.querySchemaModel;
-    this.schema = this._generateSchema();
-    this._initBinds();
-  },
-
-  _initBinds: function () {
-    this.bind('change', this._onChange, this);
-  },
-
   _onChange: function () {
     this._styleModel.set(this.changed);
-  },
-
-  _generateSchema: function () {
-    var self = this;
-    return _.reduce(this.attributes, function (memo, value, key) {
-      var d = StyleFormComponents[key];
-      if (d) {
-        memo[key] = d(self._querySchemaModel, self._styleModel.get('type'));
-      }
-      return memo;
-    }, {});
   }
 
 });
